+++ src/siri/db/ffile.c
@@ -121,7 +121,7 @@ siridb_ffile_t * siridb_ffile_new(
             return NULL;
         }
 
-        if ((ffile->size = ftello(ffile->fp)) >= sizeof(uint32_t))
+        if ((ffile->size = ftello(ffile->fp)) >= (off_t) sizeof(uint32_t))
         {
             ffile->free_space = 0;
             if (    fseeko(ffile->fp, -(long int) sizeof(uint32_t), SEEK_END) ||
